+++ src/app/store.js
@@ -1,5 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit';
-import gameReducer from '../reducers/recipeSlice';
+import gameReducer from '../reducers/gameSlice';
 
 const store = configureStore({
   reducer: {
