+++ src/views/Game-info.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import GameInfoList from '../containers/GameInfoList';
+
+function GameInfo() {
+  return (
+    <div className="recipes">
+      <GameInfoList />
+    </div>
+  );
+}
+
+export default GameInfo;
