--- src/views/Recipes.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import RecipeList from '../containers/RecipeList';
-
-function Recipes() {
-  return (
-    <div className="recipes">
-      <RecipeList />
-    </div>
-  );
-}
-
-export default Recipes;
